pyproject.toml
[project]
name = "influencer-factory"
version = "0.1.0"
description = "AI virtual influencer content pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "jinja2>=3.1",
    "tenacity>=8.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "anthropic>=0.49",
    "openai>=1.30",
    "pillow>=10.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "streamlit>=1.33",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
